feat: Change phoneNumber from Long to String in Contact model and related components
- Updated `Contact` and `ContactEntity` models to use `String` for `phoneNumber`.
- Modified relevant logic in `ContactEditScreen` and `AddContactScreen` to handle `phoneNumber` as a `String`.
- Incremented `ConnectifyDatabase` version to 2 to reflect schema change.

+++ app/src/main/java/com/example/connectify/presentation/screens/contact/AddContactScreen.kt
@@ -91,7 +91,7 @@ fun AddContactScreen(
                     Contact(
                         name = name,
                         email = email,
-                        phoneNumber = phoneNumber.toLong(),
+                        phoneNumber = phoneNumber,
                         imageUrl = imageUri?.toString()
                     )
                 )
